refactored uploadPlugin API requests to use new fetch functions
