+++ src/js/engagement_view/src/components/uploadPlugin/utils/uploadPluginApiRequests.tsx
@@ -1,60 +1,61 @@
 import { PluginPayload } from "./uploadPluginTypes";
-import {getModelPluginEdge} from "../../../services/getApiURLs"
+import {getModelPluginEdge} from "../../../services/getApiURLs";
 
-export const getPluginList = async () => {
-    console.log("getModelPluginEdge", getModelPluginEdge())
-    const res = await fetch(`${getModelPluginEdge()}listModelPlugins`, 
-        {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            credentials: 'include',
-        }
-    );
+// export const getPluginList = async () => {
+//     // console.log("getModelPluginEdge", getModelPluginEdge())
+//     const res = await fetch(`${getModelPluginEdge()}listModelPlugins`, 
+//         {
+//             method: 'POST',
+//             headers: {
+//                 'Content-Type': 'application/json',
+//             },
+//             credentials: 'include',
+//         }
+//     );
     
-    console.log("res", res);
+//     // console.log("res", res);
 
-    const body = await res.json() as any;
+//     const body = await res.json() as any;
     
-    console.log("body", body)
-    try { 
-        let pluginList: string[] = body.success.plugin_list;
+//     // console.log("body", body);
 
-        return pluginList
-    } catch (e) {
-        console.warn("Error retrieving pluginList", e);
-        return [];
-    }
-}
+//     try { 
+//         let pluginList: string[] = body.success.plugin_list;
 
-export const deletePlugin = async ( pluginName: string ): Promise <boolean> => {
-    const res = await fetch(`${getModelPluginEdge()}deleteModelPlugin`, 
-        {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            credentials: 'include',
-            body: JSON.stringify({plugins_to_delete: [pluginName]})
-        }
-    );
-    await res.json();
-    return true;
-};
+//         return pluginList
+//     } catch (e) {
+//         console.warn("Error retrieving pluginList", e);
+//         return [];
+//     }
+// }
 
-export const uploadFilesToDgraph = async (payload: PluginPayload ): Promise<boolean> => {
-    const res = await fetch(`${getModelPluginEdge()}deploy`, 
-        {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            credentials: 'include',
-            body: JSON.stringify(payload)
-        }
-    );
-    const body = await res.json();
+// export const deletePlugin = async ( pluginName: string ): Promise <boolean> => {
+//     const res = await fetch(`${getModelPluginEdge()}deleteModelPlugin`, 
+//         {
+//             method: 'POST',
+//             headers: {
+//                 'Content-Type': 'application/json',
+//             },
+//             credentials: 'include',
+//             body: JSON.stringify({plugins_to_delete: [pluginName]})
+//         }
+//     );
+//     await res.json();
+//     return true;
+// };
 
-    return body.success.Success;
-};
+// export const uploadFilesToDgraph = async (payload: PluginPayload ): Promise<boolean> => {
+//     const res = await fetch(`${getModelPluginEdge()}deploy`, 
+//         {
+//             method: 'POST',
+//             headers: {
+//                 'Content-Type': 'application/json',
+//             },
+//             credentials: 'include',
+//             body: JSON.stringify(payload)
+//         }
+//     );
+//     const body = await res.json();
+
+//     return body.success.Success;
+// };
