+++ src/js/engagement_view/src/components/uploadPlugin/utils/uploadPlugins.tsx
@@ -5,7 +5,7 @@ import Button from "@material-ui/core/Button";
 import CloudUploadIcon from '@material-ui/icons/CloudUpload';
 
 import { Event, UploadFormState, DirectoryUpload, MessageProps} from "./uploadPluginTypes"
-import { uploadFilesToDgraph } from "./uploadPluginApiRequests";
+import { uploadFilesToDgraph } from "../../../services/uploadPlugin/uploadFilesToDgraph";
 import { useStyles } from "../uploadPluginStyles";
 
 const readFile = async (file: Blob): Promise <string> => {
